third_party/WebKit/LayoutTests/http/tests/websocket/cookie-ws-to-ws-expected.txt

Test send cookies to a WebSocket handshake which is set by another WebSocket handshake.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

A WebSocket connection to ws://127.0.0.1:8880/set-cookie is opened.
A WebSocket connection to ws://127.0.0.1:8880/set-cookie is closed.
A WebSocket connection to ws://127.0.0.1:8880/echo-cookie is opened.
A WebSocket connection to ws://127.0.0.1:8880/echo-cookie is closed.
PASS cookie is "WK-WebSocket-test-domain-pass=1; WK-WebSocket-test-path-pass=1; WK-WebSocket-test=1"
PASS successfullyParsed is true

TEST COMPLETE
